Extract rudimentary type info for debugged vars
This will only work for base types, we only check the size of the
variable and mask the probed u64 to the appropriate width.

+++ src/debugee.zig
@@ -11,7 +11,7 @@ var my_global_var_2: GlobalVarTracker = .a;
 
 fn doPrintLoop(val: usize) void {
     const x: i32 = 4 + @as(i32, @intCast(val));
-    const y: i32 = 5 + @as(i32, @intCast(val));
+    const y: u8 = 5 + @as(u8, @intCast(val));
     my_global_var_pog += 1;
     std.debug.print("loop iter {d} {d} {d} {d} {any}\n", .{ x, y, val, my_global_var_pog, my_global_var_2 });
 }
